Fixed some trkilat

+++ src/extra/utils.c
@@ -20,5 +20,5 @@ float normalize(float angle)
 
 float distance_point2point(float x1, float y1, float x2, float y2)
 {
-	return sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
+	return sqrtf((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
 }
